Fix float!=bfloat16: force all model params to bf16 after .cuda()
use_cpu_initialization=True does not guarantee every submodule (LayerNorm,
embeddings, biases) respects params_dtype=bfloat16. Adding .bfloat16() after
.cuda() casts every parameter and buffer unconditionally, eliminating the
float != c10::BFloat16 dtype mismatch in TP linear layers.

+++ Megatron.py
@@ -60,7 +60,7 @@ def _build_model(model_cfg: dict, seq_len: int) -> GPTModel:
         pre_process=mpu.is_pipeline_first_stage(),
         post_process=mpu.is_pipeline_last_stage(),
     )
-    return model.cuda()
+    return model.cuda().bfloat16()
 
 
 class _DataIterator:
